@@ -1,27 +1,27 @@
 {
-  "name": "agricom",
+  "name": "client",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "axios": "^0.21.4",
-    "dayjs": "^1.10.7",
-    "jwt-decode": "^3.1.2",
-    "jwt-encode": "^1.0.1",
-    "miragejs": "^0.1.41",
-    "mockman-js": "^1.0.5",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-router-dom": "^6.2.2",
-    "react-scripts": "4.0.3",
-    "react-toastify": "^8.2.0",
-    "uuid": "^8.3.2",
-    "web-vitals": "^0.2.4"
+    "@testing-library/jest-dom": "^5.14.1",
+    "@testing-library/react": "^13.0.0",
+    "@testing-library/user-event": "^13.2.1",
+    "react": "^18.1.0",
+    "react-dom": "^18.1.0",
+    "react-scripts": "5.0.1",
+    "web-vitals": "^2.1.0"
   },
   "scripts": {
     "start": "react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test",
     "eject": "react-scripts eject"
+  },
+  "eslintConfig": {
+    "extends": [
+      "react-app",
+      "react-app/jest"
+    ]
   },
   "browserslist": {
     "production": [
@@ -34,10 +34,5 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
-  },
-  "devDependencies": {
-    "eslint": "^7.32.0",
-    "eslint-plugin-react": "^7.29.4",
-    "eslint-plugin-react-hooks": "^4.4.0"
   }
 }
