@@ -1,15 +1,23 @@
-import "./App.css";
-import "./styles/utilities/variables.css";
-import { PageRoutes } from "./routes/PageRoutes";
-import "react-toastify/dist/ReactToastify.css";
-
-import { ToastContainer } from "react-toastify";
+import logo from './logo.svg';
+import './App.css';
 
 function App() {
   return (
     <div className="App">
-      <ToastContainer autoClose={1500} />
-      <PageRoutes />
+      <header className="App-header">
+        <img src={logo} className="App-logo" alt="logo" />
+        <p>
+          Edit <code>src/App.js</code> and save to reload.
+        </p>
+        <a
+          className="App-link"
+          href="https://reactjs.org"
+          target="_blank"
+          rel="noopener noreferrer"
+        >
+          Learn React
+        </a>
+      </header>
     </div>
   );
 }
